requirements.txt
pandas
python-dotenv
langchain
langchain-community
langchain-google-genai
faker
chromadb
langchain-chroma